Helper functions moved into binaries README minor improvements
