+++ contracts/.env.template
@@ -4,6 +4,7 @@ PRIVATE_KEY=YOUR_PRIVATE_KEY
 # The RPC endpoint to use for deploying the contracts and reading the state
 # NOTE: this is a public endpoint, so it may be slow or unreliable
 SEPOLIA_RPC_URL=https://rpc2.sepolia.org
+MAINNET_RPC_URL=tofill
 
 # Add your Etherscan API key here to verify the contracts on Etherscan
 ETHERSCAN_API_KEY=YOUR_ETHERSCAN_API_KEY
