+++ contracts/README.md
@@ -6,7 +6,8 @@ This repository contains all the smart contracts for the Nouns Anonymous Voting 
 
 To redeploy the contracts, first copy the `.env.template` file to `.env` and fill in the values.
 
-Make sure you also have `forge` installed. You can read more about Foundry and Forge [here](https://book.getfoundry.sh/getting-started/installation).
+Make sure you also have `forge` installed. You can read more about Foundry and
+Forge [here](https://book.getfoundry.sh/getting-started/installation).
 
 ## Deploying
 
@@ -29,7 +30,8 @@ source .env
 
 ### Sepolia Network
 
-**Note:** Make sure you have the correct `SEPOLIA_RPC_URL` set in your `.env` file, as well as the `PRIVATE_KEY` maps to an account that has enough funds to deploy the contracts.
+**Note:** Make sure you have the correct `SEPOLIA_RPC_URL` set in your `.env` file, as well as the `PRIVATE_KEY` maps to
+an account that has enough funds to deploy the contracts.
 
 Run the following commands to deploy the contracts:
 
@@ -37,4 +39,12 @@ Run the following commands to deploy the contracts:
 source .env
 
 forge script script/2_NounsVoting.s.sol:NounsVotingDeployScript --rpc-url $SEPOLIA_RPC_URL --broadcast --verify -vvvv
+```
+
+### Mainnet
+
+To deploy the contract to mainnet, populate the `.env` file with the correct values and run the following command:
+
+```bash
+forge script script/3_NounsMainnetVoting.s.sol:NounsMainnetVotingDeployScript --rpc-url $MAINNET_RPC_URL --broadcast --verify -vvvv
 ```
