+++ nouns-cli/Cargo.toml
@@ -7,6 +7,7 @@ authors = ["AZKR Team"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+dotenv = "0.15"
 eyre = "0.6"
 ethers = { version = "2.0", features = ["rustls"] }
 tokio = { version = "1.0", features = ["rt", "rt-multi-thread", "macros"] }
@@ -14,5 +15,3 @@ clap = { version = "4.3.3", features = ["derive", "cargo", "env"] }
 
 nouns-protocol = { path = "../nouns-protocol", features = ["mock-prover"] }
 
-[dev-dependencies]
-dotenv = "0.15"
