+++ nouns-cli/src/lib.rs
@@ -1,0 +1,11 @@
+pub(crate) use ethers::prelude::U256 as EthersU256;
+
+pub use ethereum::contract_interactions::{
+    mine_blocks_until, obtain_token_ids_to_vote, NounsVoting,
+};
+pub use ethereum::setup_connection;
+pub use ethereum::setup_env_parameters;
+
+pub mod cli;
+pub mod ethereum;
+mod parsers;
